+++ release_notes.md
@@ -4,4 +4,5 @@
 - My change description (#PR)
 -->
 - Update Python Worker Version to [4.23.0](https://github.com/Azure/azure-functions-python-worker/releases/tag/4.23.0)
+- Updated `Microsoft.Azure.Functions.DotNetIsolatedNativeHost` version to 1.0.5 (#9753)
 - Add function grouping information (https://github.com/Azure/azure-functions-host/pull/9735)